Remove the "deployed before metallb" test
